add some methods to DAO User, Order, Address, Car + long id to Car, Address

+++ Server/src/main/java/ua/artcode/taxi/dao/GenericDao.java
@@ -13,7 +13,7 @@ public interface GenericDao<T> {
 
     boolean delete(T el);
 
-    T findById(int id);
+    T findById(long id);
 
     List<T> getAll(int offset, int length);
